Improved merging of null-values
YAML returns 'None' when a key is parsed without a value. This is
properly handled by a95eaa35f5c48417420fe37472d8148fdbf5ddd5, except for
the recursive dict merger, which shouldn't overwrite existing data
structures when None is encountered. The solution is simply to provide
appropriate policy entries for when the second value is a NoneType.

Signed-off-by: martin f. krafft <madduck@madduck.net>