Limit top() output to minion if ID specified
Salt expects the top data to be limited to those states applicable to
the very minion being processed. Therefore, the interface needs to be
extended to accept minion_id, and if that is specified, then only the
minion-specific states are returned; else, the entire top data are
returned, as before (which the CLI still does).

Thanks to David Anderson for figuring this out!

Signed-off-by: martin f. krafft <madduck@madduck.net>