Define argh-compatible func/cmd and entry points for setuptools
The entry points are used by ``console_scripts``, and are simple
functions leveraging ``argh`` for option/argument parsing and command
dispatching.

This initial commit simply defines the new functions but does not
update Reclass to use them.

Note that the salt/ansible adapter commands are simple place-holders
for the moment.