Move information about invocation modes up to CLI
I'm writing an adapter for Puppet and had to discover that I hard-coded
the assumption that reclass and its adapters would always be invoked in
either of two modes (inventory & nodeinfo). This is true for Salt and
Ansible, and it makes sense for the main CLI too, but Puppet has no
concept of inventory (since it's pull-based, unlike the others), and
thus only has one mode.

In order to support this, I've moved all assumptions about how to invoke
reclass into the adapters and the CLI script, such that it should be
possible now to write a Puppet adapter that does not need to jump
through hoops to avoid functionality it does not need.

Signed-off-by: martin f. krafft <madduck@madduck.net>